python-api/requirements.txt
flask>=3.0
pillow>=10.0
torch>=2.0
torchvision>=0.15
